refactor: 파일 분리 및 enrollPlayers(), inputCarNames(), verifyCarName() 역할 수정
