+++ src/main/java/domain/Main.java
@@ -6,110 +6,8 @@ import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
-        init();
-    }
-
-    private static void init() {
-        List<Car> carsList = inputCarNames();
-        startRace(carsList, getTurnNumber());
-        printWinner(carsList);
-    }
-
-    private static void printWinner(List<Car> carsList) {
-        List<String> winnerList = getWinner(carsList);
-        for(int i = 0; i < winnerList.size(); i++) {
-            System.out.printf("%s", winnerList.get(i));
-            if(i != winnerList.size()-1) {
-                System.out.printf(", ");
-            }
-        }
-        System.out.printf("님이 최종 우승했습니다.\n");
-    }
-
-    private static List<String> getWinner(List<Car> carsList) {
-        int maxScore = getMaxScore(carsList);
-        List<String> winnerList = new ArrayList<String>();
-        for(Car oneCar: carsList) {
-            if (oneCar.getPosition() == maxScore) {
-                winnerList.add(oneCar.getName());
-            }
-        }
-        return winnerList;
-    }
-
-    private static int getMaxScore(List<Car> carsList) {
-        int maxScore = 0;
-        for(Car oneCar: carsList) {
-            int thisCarPosition = oneCar.getPosition();
-            if (thisCarPosition > maxScore) {
-                maxScore = thisCarPosition;
-            }
-        }
-        return maxScore;
-    }
-
-    private static void startRace(List<Car> carsList, int turnNumber) {
-        System.out.println("실행 결과");
-        for(int i = 0; i < turnNumber; i++) {
-            playOneTurn(carsList);
-        }
-    }
-
-    private static void playOneTurn(List<Car> carsList) {
-        for(Car oneCar : carsList) {
-            oneCar.playTurn();
-        }
-        System.out.println();
-    }
-
-    private static int getTurnNumber() {
-        int turnNumber = 0;
-        Scanner scan = new Scanner(System.in);
-        while(true) {
-            System.out.println("시도할 횟수는 몇 회인가요?");
-            turnNumber = scan.nextInt();
-            if(turnNumber > 0 ) {
-                break;
-            }
-            System.out.println("다시 입력해 주세요.");
-        }
-        return turnNumber;
-    }
-
-    private static List<Car> inputCarNames() {
-        String carNames;
-        List<Car> carsList;
-        Scanner scan = new Scanner(System.in);
-        while (true) {
-            System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
-            carNames = scan.nextLine();
-            carsList = enrollPlayers(carNames);
-            if (carsList != null) {
-                return carsList;
-            }
-            System.out.println("다시 입력해 주세요.");
-        }
-    }
-    
-    
-
-    private static List<Car> enrollPlayers(String carNames) {
-        String[] carArray = carNames.split(",");
-        List<Car> carsList = new ArrayList<Car>();
-        for (String oneCar : carArray) {
-            if(!verifyCarName(oneCar)) {
-                return null;
-            }
-            carsList.add(new Car(oneCar));
-        }
-        return carsList;
-    }
-
-    private static boolean verifyCarName(String carName) {
-        if(carName.length() > 5 || carName.length() == 0) {
-            return false;
-        }
-        return true;
+        Game game = new Game();
+        game.startGame();
     }
 }
